developing (last serial task & docs & .fzz to do)

+++ Assignment-03/esp8266/espmain/Light_sensor.h
@@ -5,8 +5,7 @@
 class Light_sensor{ 
 public:
   Light_sensor(int pin);
-  //void lightSensorSetup();
-  int getLightLevel();    
+  double getLightLevel();    
 private:
   int pin;  
 };
